aabb margin experiments spinner benchmark
