+++ src/aabb.h
@@ -51,14 +51,6 @@ static inline bool b2EnlargeAABB( b2AABB* a, b2AABB b )
 /// Do a and b overlap
 static inline bool b2AABB_Overlaps( b2AABB a, b2AABB b )
 {
-	b2Vec2 d1 = { b.lowerBound.x - a.upperBound.x, b.lowerBound.y - a.upperBound.y };
-	b2Vec2 d2 = { a.lowerBound.x - b.upperBound.x, a.lowerBound.y - b.upperBound.y };
-
-	if ( d1.x > 0.0f || d1.y > 0.0f )
-		return false;
-
-	if ( d2.x > 0.0f || d2.y > 0.0f )
-		return false;
-
-	return true;
+	return !( b.lowerBound.x > a.upperBound.x || b.lowerBound.y > a.upperBound.y || a.lowerBound.x > b.upperBound.x ||
+			  a.lowerBound.y > b.upperBound.y );
 }
