+++ src/core.h
@@ -162,7 +162,10 @@ extern float b2_lengthUnitsPerMeter;
 // to move by a small amount without triggering a tree adjustment.
 // This is in meters.
 // @warning modifying this can have a significant impact on performance
-#define b2_aabbMargin ( 0.1f * b2_lengthUnitsPerMeter )
+#define b2_aabbMargin ( 0.02f * b2_lengthUnitsPerMeter )
+
+// todo testing
+#define b2_aabbVelocityScale 0.0f
 
 // The time that a body must be still before it will go to sleep. In seconds.
 #define b2_timeToSleep 0.5f
